Corrección de constructores y pruebas unitarias

+++ TorneoTCG/src/main/java/com/example/TorneoTCG/dto/JugadorDTO.java
@@ -13,12 +13,10 @@ import lombok.NoArgsConstructor;
 @Builder
 
 public class JugadorDTO {
-    private Integer id;
+    private Long id;
     private String nombre;
     private String apellido;
     private String email;
     private List<String> nombreMazos;
-    
+
 }
-
-
